server/requirements.txt
Flask==2.2.2
Flask_Cors==3.0.10
googlemaps==4.10.0
python-dotenv==1.0.0
Werkzeug==2.2.2
